requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pandas
numpy
scikit-learn
xgboost
joblib       
python-dotenv
streamlit